home_page: consolidate sitemaps into config/
Move StaticViewSitemap and CourseSitemap into config/sitemaps.py, the
project's composition root, and delete the per-app sitemap modules. This
removes the cross-app coupling where base and content_engine reversed
student_interface URL names, and co-locates all sitemaps in one place.
